+++ FluxSolution/BaseLibrary/Source/Types/Stopwatch/GetMilliseconds.em.cs
@@ -4,6 +4,6 @@ namespace Flux
 	{
 		/// <summary>Compute how many milliseconds has elapsed.</summary>
 		public static double GetTotalMilliseconds(this System.Diagnostics.Stopwatch source)
-			=> (double)(source ?? throw new System.ArgumentNullException(nameof(source))).ElapsedTicks / (double)System.Diagnostics.Stopwatch.Frequency * 1000.0;
+			=> (source ?? throw new System.ArgumentNullException(nameof(source))).ElapsedTicks / (double)System.Diagnostics.Stopwatch.Frequency * 1000.0;
 	}
 }
